Out-of-vocab words filtered with NgramModelVocabulary, not NgramCounter.
