+++ nltk/model/util.py
@@ -1,0 +1,11 @@
+# Natural Language Toolkit: Language Model Unit Tests
+#
+# Copyright (C) 2001-2016 NLTK Project
+# Author: Ilia Kurenkov <ilia.kurenkov@gmail.com>
+# URL: <http://nltk.org/>
+# For license information, see LICENSE.TXT
+
+
+def mask_oov_words_in_corpus(corpus, ngram_vocab):
+    """Replace all out-of-vocabulary words in a corpus label for unknowns."""
+    return [list(map(ngram_vocab.mask_oov, sent)) for sent in corpus]
